social_media_api/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.0.1
django-filter==23.3
django-cors-headers==4.3.1
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
dj-database-url==2.0.0
python-decouple==3.8